Update Economic.md
In the spirit of contributing in extremely small ways to the Digital Knowledge Loop a la Wikipedia, I've made some suggestions for minor grammatical corrections here and in other sections. Thank you for contributing this work, it is one of the most insightful and inspiring books I've read in a long time!